disabled ability to delete friends events

+++ src/components/events/EventCard.js
@@ -8,7 +8,9 @@ import { Card, Button } from 'react-bootstrap';
 
 export class NextEventCard extends Component {
 
-    render() {   
+    render() {
+        const userId = localStorage.getItem("userId")
+        if (Number(userId) === this.props.event.userId) {
             return (
                 <div className="eventCard card nextEvent">
                     <Card>
@@ -24,27 +26,60 @@ export class NextEventCard extends Component {
                         </div>
                     </Card>
                 </div>
-            ); 
-        } 
-    }
-    export class RegularEventCard extends Component {
-
-        render() {   
-                return (
-                    <div className="eventCard card">
-                        <Card>
-                            <div className="card-content">
-                                <Card.Body>
-                                    <h2><span className="card-eventname">{this.props.event.eventName}</span></h2>
-                                    <p>Date: {this.props.event.eventDate}</p>
-                                    <p>Location: {this.props.event.location}</p>
-                                    <Button variant="primary" type="button" className="events-button" onClick={() => this.props.deleteEvent(this.props.event.id)}>Delete</Button>
-                                    <Button variant="primary" type="button" className="events-button"
-                                        onClick={() => { this.props.history.push(`/events/${this.props.event.id}/edit`) }}>Edit</Button>
-                                </Card.Body>
-                            </div>
-                        </Card>
-                    </div>
-                ); 
-            } 
+            );
+        } else {
+            return (
+                <div className="eventCard card nextEvent">
+                    <Card>
+                        <div className="card-contentnext nextEvent">
+                            <Card.Body>
+                                <h2><span className="card-eventname">{this.props.event.eventName}</span></h2>
+                                <p>Date: {this.props.event.eventDate}</p>
+                                <p>Location: {this.props.event.location}</p>
+                            </Card.Body>
+                        </div>
+                    </Card>
+                </div>
+            );
         }
+    }
+}
+export class RegularEventCard extends Component {
+
+    render() {
+        const userId = localStorage.getItem("userId")
+        if (Number(userId) === this.props.event.userId) {
+            return (
+                <div className="eventCard card">
+                    <Card>
+                        <div className="card-content">
+                            <Card.Body>
+                                <h2><span className="card-eventname">{this.props.event.eventName}</span></h2>
+                                <p>Date: {this.props.event.eventDate}</p>
+                                <p>Location: {this.props.event.location}</p>
+                                <Button variant="primary" type="button" className="events-button" onClick={() => this.props.deleteEvent(this.props.event.id)}>Delete</Button>
+                                <Button variant="primary" type="button" className="events-button"
+                                    onClick={() => { this.props.history.push(`/events/${this.props.event.id}/edit`) }}>Edit</Button>
+                            </Card.Body>
+                        </div>
+                    </Card>
+                </div>
+            );
+        } else {
+            return (
+                <div className="eventCard card">
+                    <Card>
+                        <div className="card-content">
+                            <Card.Body>
+                                <h2><span className="card-eventname">{this.props.event.eventName}</span></h2>
+                                <p>Date: {this.props.event.eventDate}</p>
+                                <p>Location: {this.props.event.location}</p>
+                            </Card.Body>
+                        </div>
+                    </Card>
+                </div>
+
+            )
+        }
+    }
+}
